requirements.txt
Django==1.11.2
Faker==0.7.17
django-filter==1.0.4
djangorestframework==3.6.3
factory-boy==2.8.1
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
